+++ lib/src/annotations/import.dart
@@ -2,6 +2,7 @@
 library jaguar.src.annotations;
 
 import 'dart:async';
+import 'dart:convert';
 
 import 'package:jaguar/jaguar.dart';